Add API endpoint to create a spotted Treasure by User
This endpoint is a very simple implementation that illustrates
the app's basic architecture, gems and practices used for future app's
development. Thus it lacks many features for now (like there is a fixed
"current user" in the controller since there's not auth implemented yet).

Thus it was used only to give an overview of the approach to project's
implementation and is a subject to future extension.